# openapi.yaml
openapi: 3.0.3
info:
  title: Strategic Planning Platform API
  description: |
    AI-Powered Strategic Planning Platform with GraphRAG validation.
    
    ## Features
    
    * **4-Phase PRD Creation Workflow**: From concept to comprehensive document
    * **GraphRAG Validation**: <2% hallucination rate through three-tier validation
    * **PydanticAI Agents**: Intelligent PRD processing and task generation
    * **GitHub Integration**: Automated project setup and task management
    * **Enterprise Security**: JWT auth, RBAC, and comprehensive audit logging
    
    ## Performance Targets
    
    * API Response Time: <200ms P95
    * Document Generation: <60 seconds
    * Concurrent Users: 100+ with stable performance
    * Uptime: 99.9% SLA
  version: 1.0.0
  contact:
    name: Strategic Planning Platform Team
    email: support@strategicplanning.ai
  license:
    name: Proprietary
servers:
  - url: http://localhost:8000
    description: Development server
  - url: https://api.strategicplanning.ai
    description: Production server
security:
  - BearerAuth: []
paths:
...
  /api/v1/validation/validate-content:
    post:
      tags:
        - GraphRAG Validation
      summary: Validate content
      description: Validate content using GraphRAG three-tier validation system
      operationId: validateContent
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidationRequest'
      responses:
        '200':
          description: Validation completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationResponse'
        '400':
          description: Invalid content
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
  /validate-content:
    post:
      tags:
        - GraphRAG Validation
      summary: Validate content (alias)
      operationId: validateContentAlias
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ValidationRequest'
      responses:
        '200':
          description: Validation completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ValidationResponse'
  /validate-batch:
    post:
      tags:
        - GraphRAG Validation
      summary: Validate content batch (alias)
      operationId: validateBatchAlias
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        '200':
          description: Batch validation completed
  /confidence-score/{prd_id}:
    get:
      tags:
        - GraphRAG Validation
      summary: Get confidence score (alias)
      operationId: getConfidenceScoreAlias
      parameters:
        - name: prd_id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Confidence score retrieved
          content:
            application/json:
              schema:
                type: object
                properties:
                  confidence:
                    type: number
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      required:
        - error
        - status_code
        - timestamp
      properties:
        error:
          type: string
          description: Error message
        status_code:
          type: integer
          description: HTTP status code
        timestamp:
          type: string
          format: date-time
          description: Error timestamp
        path:
          type: string
          description: Request path
        request_id:
          type: string
          description: Request correlation ID
...
    ValidationRequest:
      type: object
      required:
        - content
      properties:
        content:
          type: string
          minLength: 10
          maxLength: 5000
          description: Content to validate
        context:
          type: object
          description: Validation context
        section_type:
          type: string
          description: Type of content section
        project_id:
          type: string
          description: Associated project ID
    ValidationResponse:
      type: object
      required:
        - validation_id
        - confidence
        - passes_threshold
        - entity_validation
        - community_validation
        - global_validation
        - requires_human_review
        - timestamp
      properties:
        validation_id:
          type: string
          description: Validation result ID
        confidence:
          type: number
          minimum: 0
          maximum: 1
          description: Overall confidence score
        passes_threshold:
          type: boolean
          description: Meets validation threshold
        entity_validation:
          type: object
          description: Entity-level validation results
        community_validation:
          type: object
          description: Community-level validation results
        global_validation:
          type: object
          description: Global-level validation results
        corrections:
          type: array
          items:
            type: string
          description: Suggested corrections
        requires_human_review:
          type: boolean
          description: Requires manual review
        timestamp:
          type: string
          format: date-time
          description: Validation timestamp
...